+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Blunder - episode 3".
 - Tests for "CGX formatter".
 - Tests for "Genome sequencing".
+- Tests for "Roller coaster".
 
 ## [1.3.0] - 2022-03-02
 ### Added